backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
python-dotenv==1.0.1
PyGithub==2.3.0
transformers==4.40.0
torch==2.3.0
faiss-cpu==1.8.0
numpy==1.26.4
rank-bm25==0.2.2
tqdm==4.66.4